split root project into root and bridgescorekeeper
