+++ bridgescorekeeper/src/test/scala/com/github/thebridsk/bridge/test/manual/InspectJars.scala
