+++ bridgescorekeeper/src/test/scala/com/github/thebridsk/bridge/test/pages/HelpPage.scala
