+++ bridgescorekeeper/src/test/scala/com/github/thebridsk/bridge/test/selenium/integrationtest/HelpTest.scala
